requirements.txt
pdfplumber==0.11.0
opencc-python-reimplemented==0.1.7
reportlab==4.1.0
PyPDF2==3.0.1
